// src/components/Link/Link.d.ts
export interface LinkProps {
    href: string;
    text: string;
    external?: boolean;
    underline?: boolean;
    className?: string;
}
declare function Link({ href, text, external, underline, className, }: LinkProps): import("react/jsx-runtime").JSX.Element;
export default Link;
//# sourceMappingURL=Link.d.ts.map